Move game key handling to game window
