+++ src/app/app.rs
@@ -49,36 +49,12 @@ impl App {
 
     fn on_key_event(&mut self, key: KeyEvent) {
         match (key.modifiers, key.code) {
-            // [Quit combinations]
+            // [Instant quit combinations]
             (_, KeyCode::Esc | KeyCode::Char('q'))
             | (KeyModifiers::CONTROL, KeyCode::Char('c') | KeyCode::Char('C')) => self.quit(),
-            // [Arrow navigation]
-            (_, KeyCode::Left | KeyCode::Char('h')) => self.on_left_pressed(),
-            (_, KeyCode::Down | KeyCode::Char('j')) => self.on_down_pressed(),
-            (_, KeyCode::Up | KeyCode::Char('k')) => self.on_up_pressed(),
-            (_, KeyCode::Right | KeyCode::Char('l')) => self.on_right_pressed(),
-            // [Enter]
-            (_, KeyCode::Enter) => self.on_enter_pressed(),
             // Add other key handlers here
-            _ => {}
+            _ => self.game_window.on_key_pressed(key),
         }
-    }
-
-    fn on_enter_pressed(&mut self) {
-        self.game_window.deal_cards();
-    }
-
-    fn on_up_pressed(&self) {
-        todo!("TODO: [⤴]")
-    }
-    fn on_down_pressed(&self) {
-        todo!("TODO: [↓]")
-    }
-    fn on_left_pressed(&self) {
-        todo!("TODO: [←]")
-    }
-    fn on_right_pressed(&self) {
-        todo!("TODO: [→]")
     }
 
     fn quit(&mut self) {
