+++ src/app/game_window.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 
+use crossterm::event::{KeyCode, KeyEvent};
 use ratatui::{
     Frame,
     layout::{Constraint, Layout, Margin, Position, Rect},
@@ -23,10 +24,41 @@ impl<CardStockT: ICardStock> GameWindow<CardStockT> {
         Self { game_engine }
     }
 
-    pub fn deal_cards(&mut self) {
+    fn deal_cards(&mut self) {
         self.game_engine.deal_cards();
     }
 
+    pub fn on_key_pressed(&mut self, key: KeyEvent) {
+        match (key.modifiers, key.code) {
+            // [Arrow navigation]
+            (_, KeyCode::Left | KeyCode::Char('h')) => self.on_left_pressed(),
+            (_, KeyCode::Down | KeyCode::Char('j')) => self.on_down_pressed(),
+            (_, KeyCode::Up | KeyCode::Char('k')) => self.on_up_pressed(),
+            (_, KeyCode::Right | KeyCode::Char('l')) => self.on_right_pressed(),
+            // [Enter]
+            (_, KeyCode::Enter) => self.on_enter_pressed(),
+            _ => {}
+        }
+    }
+
+    // -- Keys -- //
+    fn on_enter_pressed(&mut self) {
+        self.deal_cards();
+    }
+    fn on_up_pressed(&self) {
+        todo!("TODO: [⤴]")
+    }
+    fn on_down_pressed(&self) {
+        todo!("TODO: [↓]")
+    }
+    fn on_left_pressed(&self) {
+        todo!("TODO: [←]")
+    }
+    fn on_right_pressed(&self) {
+        todo!("TODO: [→]")
+    }
+
+    // --- Render --- //
     pub fn render_window(&self, frame: &mut Frame) {
         let areas = Layout::vertical([Constraint::Percentage(10), Constraint::Percentage(90)])
             .split(frame.area());
